DASH style support WORM Slide mode.
