+++ bannerview/src/main/java/com/zhpan/bannerview/indicator/drawer/BaseDrawer.java
@@ -1,6 +1,7 @@
 package com.zhpan.bannerview.indicator.drawer;
 
 import android.graphics.Paint;
+import android.graphics.RectF;
 
 import com.zhpan.bannerview.manager.IndicatorOptions;
 
@@ -12,17 +13,37 @@ import com.zhpan.bannerview.manager.IndicatorOptions;
  */
 public abstract class BaseDrawer implements IDrawer {
 
+    private MeasureResult mMeasureResult;
     IndicatorOptions mIndicatorOptions;
-
+    float maxWidth;
+    float minWidth;
     Paint mPaint;
-
-    MeasureResult mMeasureResult;
+    RectF mRectF;
 
     BaseDrawer(IndicatorOptions indicatorOptions) {
         this.mIndicatorOptions = indicatorOptions;
         mPaint = new Paint();
         mPaint.setAntiAlias(true);
         mMeasureResult = new MeasureResult();
+        mRectF = new RectF();
+    }
+
+    @Override
+    public MeasureResult onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
+        maxWidth = Math.max(mIndicatorOptions.getNormalIndicatorWidth(), mIndicatorOptions.getCheckedIndicatorWidth());
+        minWidth = Math.min(mIndicatorOptions.getNormalIndicatorWidth(), mIndicatorOptions.getCheckedIndicatorWidth());
+        mMeasureResult.setMeasureResult(measureWidth(), measureHeight());
+        return mMeasureResult;
+    }
+
+    protected int measureHeight() {
+        return (int) mIndicatorOptions.getSliderHeight();
+    }
+
+    private int measureWidth() {
+        int pageSize = mIndicatorOptions.getPageSize();
+        float indicatorGap = mIndicatorOptions.getIndicatorGap();
+        return (int) ((pageSize - 1) * indicatorGap + maxWidth + (pageSize - 1) * minWidth);
     }
 
     @Override
